+++ pytsite/permissions/__init__.py
@@ -10,7 +10,7 @@ __license__ = 'MIT'
 
 
 def _init():
-    define_group('app', 'app@app_name')
+    define_group('app', 'app_name')
 
 
 _init()